Make it possible to run GeoClue2 location provider in the main thread.
This only makes sense for redshift-dbus, where the main thread is already running a g_main_loop().

Changed the interface of location providers to allow this:
 - added function to query if location provider is dynamic; this can be used to determine if it's worth waiting for a location to become available
 - added function to set a callback on location changes; this should only be called from the same thread and is regarded as optional
 - clarified that utility function provider_get_location() can be used to query if a location update has happened

Currently, the geoclue2 location provider has a "hidden" option, "use-thread" that determines if it runs in a separate thread or tries to use the
main loop from the main thread. This probably should be made an internal option or detected automatically. For now, it should be set by the user
(mainly for debugging purposes).