Added interface for desktop environment's backlight settings
Supports GNOME and MATE. Main use-case is for screens where the backlight /
screen brightness cannot be set with the normal ways, e.g. some OLED screens.
In these cases, the normal backlight / screen brightness controls do nothing,
but redshift can be used for this purpose to some extent. This interface allows
redshift to be integrated to the native controls provided by the desktop
environment.

Added example script (redshift-backlight-proxy) that can use the interface
to connect the two interfaces (redshift and the DE).